prototype/server/requirements.txt
Flask==1.1.2
python-dotenv==0.16.0
Flask-Cors==1.10.3